Protect the data from the deletion of an auth

+++ src/declarations/usecases/auth/delete/usecase.ts
@@ -2,6 +2,7 @@ import { Injectable } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { AuthorizedUseCase } from "src/declarations/common/authorized_usecase";
 import {
+  UseCaseException,
   UseCaseOk,
   UseCaseResult,
 } from "src/declarations/common/usecase_result";
@@ -10,6 +11,8 @@ import { NothingModel } from "src/declarations/models/common/nothing";
 import { AuthEntity } from "src/declarations/entities/auth";
 import { UserEntity } from "src/declarations/entities/user";
 import { Repository } from "typeorm";
+import { ParticipantEntity } from "src/declarations/entities/participant";
+import { RoomEntity } from "src/declarations/entities/room";
 
 export type Params = {
   readonly accessToken: string;
@@ -23,6 +26,10 @@ export class DeleteAuthUseCase extends AuthorizedUseCase<Params, NothingModel> {
     private readonly authRepository: Repository<AuthEntity>,
     @InjectRepository(UserEntity)
     private readonly userRepository: Repository<UserEntity>,
+    @InjectRepository(ParticipantEntity)
+    private readonly participantRepository: Repository<ParticipantEntity>,
+    @InjectRepository(RoomEntity)
+    private readonly roomRepository: Repository<RoomEntity>,
   ) {
     super(authProvider);
   }
@@ -30,9 +37,23 @@ export class DeleteAuthUseCase extends AuthorizedUseCase<Params, NothingModel> {
   protected async executeWithAuth(
     id: string,
   ): Promise<UseCaseResult<NothingModel>> {
+    const rooms = await this.roomRepository.find({
+      where: {
+        ownerId: id,
+      },
+    });
+
+    if (rooms.length !== 0) {
+      return new UseCaseException(1);
+    }
+
     await this.authRepository.delete(id);
 
     await this.userRepository.delete(id);
+
+    await this.participantRepository.delete({
+      userId: id,
+    });
 
     return new UseCaseOk(null);
   }
